+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Next
+
+- Add `ExTier.lookup/1` endpoint
 ## v0.0.2
 
 - Add `ExTier.whoami/0` endpoint